+++ gluecodium/src/test/resources/smoke/enums/output/android-kotlin/com/example/smoke/EnumOptionSet.kt
@@ -8,7 +8,7 @@
 package com.example.smoke
 
 
-enum class EnumOptionSet(private val value: Int) {
+enum class EnumOptionSet(@JvmField val value: Int) {
     ONE(1),
     TWO(2),
     THREE(4);