+++ gluecodium/src/test/resources/smoke/equatable/output/android-kotlin/com/example/smoke/Equatable.kt
@@ -10,7 +10,7 @@ package com.example.smoke
 
 class Equatable {
 
-    enum class SomeEnum(private val value: Int) {
+    enum class SomeEnum(@JvmField val value: Int) {
         FOO(0),
         BAR(1);
     }