+++ gluecodium/src/test/resources/smoke/errors/output/android-kotlin/com/example/smoke/Errors.kt
@@ -11,11 +11,11 @@ import com.example.NativeBase
 
 class Errors : NativeBase {
 
-    enum class InternalErrorCode(private val value: Int) {
+    enum class InternalErrorCode(@JvmField val value: Int) {
         ERROR_NONE(0),
         ERROR_FATAL(1);
     }
-    enum class ExternalErrors(private val value: Int) {
+    enum class ExternalErrors(@JvmField val value: Int) {
         NONE(0),
         BOOM(1),
         BUST(2);