+++ gluecodium/src/test/resources/smoke/nesting/output/android-kotlin/com/example/smoke/FreeEnum.kt
@@ -8,7 +8,7 @@
 package com.example.smoke
 
 
-enum class FreeEnum(private val value: Int) {
+enum class FreeEnum(@JvmField val value: Int) {
     FOO(0),
     BAR(1);
 }