+++ gluecodium/src/test/resources/smoke/skip/output/android-kotlin/com/example/smoke/SkipEnumeratorAutoTag.kt
@@ -8,7 +8,7 @@
 package com.example.smoke
 
 
-enum class SkipEnumeratorAutoTag(private val value: Int) {
+enum class SkipEnumeratorAutoTag(@JvmField val value: Int) {
     ONE(0),
     THREE(1);
 }